requirements.txt
Flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.0.76
ultralytics==8.0.190
requests==2.31.0
pyserial==3.5
numpy==1.24.3
